task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,22 +10,22 @@
  */
 int main(void)
 {
-	unsigned long prev, next, ans;
+	unsigned long long prev, next, ans;
 	int i;
 
 	i = 1;
 	prev = 1;
 	next = 2;
-	printf("%lu, %lu, ", prev, next);
+	printf("%llu, %llu, ", prev, next);
 	while (i <= 97)
 	{
 		ans = prev + next;
 		prev = next;
 		next = ans;
 		if (i == 97)
-			printf("%lu\n", ans);
+			printf("%llu\n", ans);
 		else
-			printf("%lu, ", ans);
+			printf("%llu, ", ans);
 		i++;
 	}
 	return (0);
